+++ src/sema/module/nodes/path/ident.c
@@ -18,7 +18,7 @@ SemaValue *sema_path_resolve_ident(SemaModule *module, SemaValue *value, AstPath
     }
     if (type->kind == SEMA_TYPE_STRUCT) {
         size_t idx = keymap_get_idx(type->structure.fields_map, segment->ident);
-        if (idx != -1) {
+        if (idx != (size_t)-1) {
             keymap_at(type->structure.fields_map, idx, field);
             segment->sema.kind = SEMA_PATH_SEGMENT_STRUCT_FIELD;
             segment->sema.struct_field_idx = idx;
